qsrspoc-1/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
alembic==1.13.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
audioread==3.0.1
backoff==2.2.1
backports.functools-lru-cache==1.6.5
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
boto3==1.34.16
botocore==1.34.16
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chainlit==0.7.700
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.15.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.24.0
dill==0.3.7
distro==1.8.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
Faker==22.1.0
fastapi==0.100.1
fastapi-socketio==0.0.10
favicon==0.7.0
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.45.0
frozendict==2.4.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
gradio==4.5.0
gradio_client==0.7.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.19.4
humanfriendly==10.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipympl==0.9.3
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kubernetes==28.1.0
langsmith==0.0.79
Lazify==0.4.0
lazy_loader==0.3
librosa==0.10.1
litellm==1.17.0
llama-index==0.9.28.post2 # ==0.11.20 had conflict with nltk 3.8.1
#llama-index
llvmlite==0.41.1
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
mmh3==4.0.1
monotonic==1.6
more-itertools==10.2.0
mplcursors==0.5.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numba==0.58.1
numexpr==2.8.7
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.7.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pooch==1.8.0
posthog==3.0.2
prompt-toolkit==3.0.41
protobuf==4.25.1
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.7
pyparsing==3.1.1
pypdf==3.17.1
PyPika==0.48.9
pyprojroot==0.3.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.3.3
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.28.2
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchvision
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
trulens-eval==0.20.2
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
uptrace==1.21.0
urllib3==1.26.18
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.10
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
